Drop the Stack.Types.Runner module, Runner includes GlobalOpts
