+++ package.yaml
@@ -244,7 +244,6 @@ library:
   - Stack.Types.Package
   - Stack.Types.PackageName
   - Stack.Types.Resolver
-  - Stack.Types.Runner
   - Stack.Types.Sig
   - Stack.Types.SourceMap
   - Stack.Types.TemplateName
